feat: ayout cardetail component (#36)
* fea: add component CarDetail

* fix-react-router (#37)

* chore: layout component

* feat: Add extended car card with more information

The extended card of the vehicle with more details of the home view is added.
It is an intermediate view between the original card and the full detail version.

* feats principal layout component

* chore: Add .vscode folder gitignore file

* chore: general grid layout carDetail component

A general layout of the component is made with Grid, where each
container is identified with a different background color.
In addition, the icons were incorporated, and the image was
adjusted to 100% width and length of its container.

* chore: refactor sx to xs and layout adjust

* chore: delete grid container icons section

delete grid container icons sections, and add align prop to grid items
for center the content inside. Add seats info to icons section

* fix: typo error alarma in Typography component

* chore: changes suggest PR #36

Specify the width of the call and whatsapp buttons to 6 each, and align
their content to the center. Main data section is separated into columns,
and the oil icon is removed

* chore: remove OilIcon import from carDetail component

* chore: delete all Grid containers childs & add objectFit property image

* fix: npm audit fix 2 vulnerabilities

* chore: Separate Typography elements carDetail section

* feats:  add general layout carDetail component

* feat: add helper for number format

* test: organizing if statements to evualuate border cases test

* test: add more cases tu evualuated test

* test: delete $ simbol test variables

Co-authored-by: Rodrigo Poblete <rodrigopoblete.dev@gmail.com>
Co-authored-by: Cristofer <cristofer-dev@users.noreply.github.com>